# Conflicts:
#	docu/graphics/roadmap_publicity.drawio
#	docu/graphics/roadmap_publicity.png

+++ admin/src/graphql/deletePendingCreation.js
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag'
 
 export const deletePendingCreation = gql`
-  mutation ($id: Float!) {
+  mutation ($id: Int!) {
     deletePendingCreation(id: $id)
   }
 `
